Do not use File.Move() as a complete file move operation.
As it was already discussed in the team you should never use File.Move()
that is known to not work when source and destination partitions (drives
on Windows) are different.

The fix is not cool, it's just using "mv", but but works. (We need working
solution on Windows everywhere later anyways.)

+++ src/Xamarin.Android.Tools.BootstrapTasks/Xamarin.Android.Tools.BootstrapTasks/UnzipDirectoryChildren.cs
@@ -109,8 +109,11 @@ namespace Xamarin.Android.Tools.BootstrapTasks {
 						: Path.Combine (destinationFolder, relativeDestDir);
 					Directory.CreateDirectory (destDir);
 					var dest    = Path.Combine (destDir, name);
-					Log.LogMessage (MessageImportance.Low, "mv '{0}' '{1}'", fse, dest);
-					File.Move (fse, dest);
+					Log.LogMessage (MessageImportance.Low, $"mv '{fse}' '{dest}'");
+					if (Directory.Exists (fse))
+						Process.Start ("/bin/mv", $@"""{fse}"" ""{dest}""").WaitForExit ();
+					else
+						File.Move (fse, dest);
 				}
 			}
 		}
